+++ src/utils/osu/misc_format.rs
@@ -52,6 +52,6 @@ pub fn format_beatmap_link(beatmap_id: i64, beatmapset_id: i64, mode: &str) -> S
     format!("https://osu.ppy.sh/beatmapsets/{beatmapset_id}#{mode}/{beatmap_id}")
 }
 
-pub fn format_user_link(user_id: u32) -> String {
+pub fn format_user_link(user_id: i64) -> String {
     format!("https://osu.ppy.sh/users/{}", user_id)
 }
